backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==2.3.3
flask-cors==6.0.0
geojson==3.1.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
packaging==25.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
shapely==2.0.3
Werkzeug==3.1.3
sqlalchemy
geoalchemy2
requests
